+++ InterplanetarySC/OperationSwitch.h
@@ -5,8 +5,8 @@
 class OperationSwitch{
 public:
 	OperationSwitch (DWORD x,DWORD y);
-	void drawSwitchON(SURFHANDLE tgt,SURFHANDLE src,double dx,double dy);
-	void drawSwitchOFF(SURFHANDLE tgt,SURFHANDLE src,double dx,double dy);
+	void drawSwitchON(SURFHANDLE tgt,SURFHANDLE src,DWORD dx,DWORD dy);
+	void drawSwitchOFF(SURFHANDLE tgt,SURFHANDLE src,DWORD dx,DWORD dy);
 
 private:
 	//OPERATION_MODE *om;
